Updated for fling API V2! Cleaned up code a bit.
The major change is that the methods have been updated to work with the
fling API V2 (the current version). This API version has some fairly
major changes such as the fact that you can no longer get a list of
'potential receivers'. When you do send a fling, the latitude and
longitude for each receiver is much less precise than API V1 which is
good news for security.